SuperficieBarrido terminada. Solo falta que funcione.

+++ trunk/inc/Curvas/Circunferencia.h
@@ -8,9 +8,9 @@
 class Circunferencia : public CurvaBSpline {
 	
 	private:
-		static std::vector<glm::vec3> mis_puntos (glm::vec3 centro, unsigned int radio);
+		static std::vector<glm::vec3> mis_puntos (glm::vec3 centro, float radio);
 	
 	public:
 		// Circunferencia paralela al plano XY
-		Circunferencia (glm::vec3 centro, unsigned int radio);
+		Circunferencia (glm::vec3 centro, float radio);
 };
